pyproject.toml
[project]
name = "florencia"
version = "0.1.0"
description = "Scrape and generate HTML with Florence's top 10 attractions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "typer>=0.15.0",
    "rich>=13.9.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.32.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
    "types-requests>=2.32.0",
    "types-pyyaml>=6.0.0",
]

[project.scripts]
florencia = "florencia.cli.main:app"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501", "N818", "UP040", "UP046", "B008", "E402", "N806"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=florencia --cov-report=term-missing"

[tool.coverage.run]
source = ["florencia"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
